This finishes the separation of variables.  We now have 3 files:   - platform.conf: Variables the framework can set at build/install time   - compset.conf: Variables that component sets can set at build/install time   - simulation ips files: platform-independent information about the         simulation "component sets" are generally something that would have a "BILDER TOPDIR" associated with it, which is always fluid.
As part of this work, the configobj.py was generalized to allow for a
tuple of filenames to be passed in and parsed at once.  This allowed for
an easy generalization of the configuration.

The refactor/test_ips.py is still failing due to the runspace init
component not creating the directory properly (I think).  Nick says he
understands it -- it didn't look obvious to me...