This now passes all of  the tests/refactor/test_ips.sh
A new multirun test directory has been created, but it is not completely
passing yet.

The Dakota files have been copied over a new tests directory has been
started, but it is not hooked up yet, nor is it ready to be because a
merge with the mainline IPS needs to take place.