Fix hang when required parameters are missing in config
If a DRIVER PORT is not-specfied you will get an error like "ERROR    Missing DRIVER ..." but ips.py never exits.

The same is if one of the require parameters for a PORT is missing, "ERROR    Error: missing required entries ..." but ips.py never exits.

This fixes the issue by instead of doing sys.exit it raise a RuntimeError in these cases